backend/requirements.txt
# VyapaarGPT Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# AI & ML
openai==1.3.7
anthropic==0.7.8
langchain==0.0.352
langchain-openai==0.0.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Audio Processing
openai-whisper==20231117
pydub==0.25.1
speech-recognition==3.10.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# API Integrations
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Data Validation & Serialization
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
loguru==0.7.2

# WhatsApp Integration
twilio==8.10.0

# Payment Integration (UPI)
razorpay==1.4.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0